Moved request and response to separate components
