+++ ui/src/components/transaction/request-body.css.ts
@@ -1,0 +1,20 @@
+import {css} from "lit";
+
+export default css`
+
+
+  .response-code {
+    font-size: 1.2em;
+  }
+
+  .contentType {
+    display: block;
+    margin-bottom: 20px;
+  }
+
+  pre {
+    max-width: calc(100vw - 135px);
+    overflow-x: auto;
+  }
+
+`
