+++ ui/src/components/transaction/response-body.css.ts
@@ -1,0 +1,16 @@
+import {css} from "lit";
+
+export default css`
+
+  .response-code {
+    font-size: 1.2em;
+  }
+
+  pre {
+    max-width: calc(100vw - 135px);
+    overflow-x: auto;
+  }
+
+
+
+`
